pyproject.toml
[tool.poetry]
name = "project-tracker"
version = "0.1.0"
description = ""
authors = ["Ryan Bui <ryan.bui@gmail.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.13"
fastapi = "^0.115.6"
python-utils-traveler = "0.0.11"
uvicorn = "^0.34.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
